+++ Magic/CHANGELOG.md
@@ -65,6 +65,7 @@
    - TODO: line of sight checks
  - Add `permissions` mob parameter, currently only works for permission-based targeting checks via the above parameters
  - Add some mob-specific disguise parameters, such as powered and ignited for creepers
+ - Added `clear_items` option to chest populator
 
 ## Fixes